Consolidate the anatomy tests into one file, and inventory tests/createcompendia
tests/test_anatomy_emapa.py held 14 tests of src/createcompendia/anatomy.py
under a name that reads as EMAPA-only, so looking in tests/createcompendia/
and concluding there was no coverage was the easy mistake -- and the one
this branch made. Merged into tests/createcompendia/test_anatomy.py,
interleaved by subject rather than appended, so the EMAPA typing tests sit
with the other typing tests and the bad-xref tests stay together. No test
bodies changed; the 14 collect and pass alongside the 16 added here.

The only anatomy tests left outside that file are the UberGraph-backed
EMAPA extraction ones in tests/pipeline/test_emapa_pipeline.py, which is
the split the layout actually intends. tests/README.md now says so, and
gains a createcompendia/ section in its Test Files inventory -- six files
that it did not mention at all.

Also name the HMDB placement path in config.yaml's build.workarounds. The
entry predates pull_hmdb() learning to reuse a pre-placed zip, so it
recorded that a manual download was needed without saying where to put it.
